ref: Split Participant and OctoParticipant.
