+++ src/main/java/org/jitsi/jicofo/conference/Participant.java
@@ -17,6 +17,7 @@
  */
 package org.jitsi.jicofo.conference;
 
+import com.google.common.collect.*;
 import org.jetbrains.annotations.*;
 import org.jitsi.impl.protocol.xmpp.*;
 import org.jitsi.jicofo.*;
@@ -46,7 +47,6 @@ import static java.time.temporal.ChronoUnit.SECONDS;
  * @author Boris Grozev
  */
 public class Participant
-        extends AbstractParticipant
 {
     /**
      * Returns the endpoint ID for a participant in the videobridge (Colibri)
@@ -60,6 +60,43 @@ public class Participant
     {
         return chatRoomMember.getName(); // XMPP MUC Nickname
     }
+
+    /**
+     * Information about Colibri channels allocated for this peer (if any).
+     */
+    private ColibriConferenceIQ colibriChannelsInfo;
+
+    /**
+     * The map of the most recently received RTP description for each Colibri
+     * content.
+     */
+    private Map<String, RtpDescriptionPacketExtension> rtpDescriptionMap;
+
+    /**
+     * List of remote source addition or removal operations that have not yet been signaled to this participant.
+     */
+    private final List<SourcesToAddOrRemove> queuedRemoteSourceChanges = new ArrayList<>();
+
+    /**
+     * Returns currently stored map of RTP description to Colibri content name.
+     * @return a <tt>Map<String,RtpDescriptionPacketExtension></tt> which maps
+     *         the RTP descriptions to the corresponding Colibri content names.
+     */
+    public Map<String, RtpDescriptionPacketExtension> getRtpDescriptionMap()
+    {
+        return rtpDescriptionMap;
+    }
+
+    /**
+     * Used to synchronize access to {@link #channelAllocator}.
+     */
+    private final Object channelAllocatorSyncRoot = new Object();
+
+    /**
+     * The {@link ParticipantChannelAllocator}, if any, which is currently
+     * allocating channels for this participant.
+     */
+    private ParticipantChannelAllocator channelAllocator = null;
 
     /**
      * The {@code BridgeSession} of which this {@code Participant} is part of.
@@ -134,12 +171,166 @@ public class Participant
             Logger parentLogger,
             JitsiMeetConferenceImpl conference)
     {
-        super(parentLogger);
-
         this.conference = conference;
         this.roomMember = roomMember;
         this.logger = parentLogger.createChildLogger(getClass().getName());
         logger.addContext("participant", getEndpointId());
+    }
+
+    /**
+     * Extracts and stores RTP description for each content type from given
+     * Jingle contents.
+     * @param jingleContents the list of Jingle content packet extension from
+     *        <tt>Participant</tt>'s answer.
+     */
+    public void setRTPDescription(List<ContentPacketExtension> jingleContents)
+    {
+        Map<String, RtpDescriptionPacketExtension> rtpDescMap = new HashMap<>();
+
+        for (ContentPacketExtension content : jingleContents)
+        {
+            RtpDescriptionPacketExtension rtpDesc
+                    = content.getFirstChildOfType(
+                    RtpDescriptionPacketExtension.class);
+
+            if (rtpDesc != null)
+            {
+                rtpDescMap.put(content.getName(), rtpDesc);
+            }
+        }
+
+        this.rtpDescriptionMap = rtpDescMap;
+    }
+
+    /**
+     * Clear the pending remote sources, indicating that they have now been signaled.
+     * @return the list of source addition or removal which have been queueed and not signaled to this participant.
+     */
+    public List<SourcesToAddOrRemove> clearQueuedRemoteSourceChanges()
+    {
+        synchronized (queuedRemoteSourceChanges)
+        {
+            List<SourcesToAddOrRemove> ret = new ArrayList<>(queuedRemoteSourceChanges);
+            queuedRemoteSourceChanges.clear();
+            return ret;
+        }
+    }
+
+    /**
+     * Gets the list of pending remote sources, without clearing them. For testing.
+     */
+    public List<SourcesToAddOrRemove> getQueuedRemoteSourceChanges()
+    {
+        synchronized (queuedRemoteSourceChanges)
+        {
+            return new ArrayList<>(queuedRemoteSourceChanges);
+        }
+    }
+
+    /**
+     * Queue a "source-add" for remote sources, to be signaled once the session is established.
+     *
+     * @param sourcesToAdd the remote sources for the "source-add".
+     */
+    public void queueRemoteSourcesToAdd(ConferenceSourceMap sourcesToAdd)
+    {
+        synchronized (queuedRemoteSourceChanges)
+        {
+            SourcesToAddOrRemove previous = Iterables.getLast(queuedRemoteSourceChanges, null);
+            if (previous != null && previous.getAction() == AddOrRemove.Add)
+            {
+                // We merge sourcesToAdd with the previous sources queued to be added to reduce the number of
+                // source-add messages that need to be sent.
+                queuedRemoteSourceChanges.remove(queuedRemoteSourceChanges.size() - 1);
+                sourcesToAdd = sourcesToAdd.copy();
+                sourcesToAdd.add(previous.getSources());
+            }
+
+            queuedRemoteSourceChanges.add(new SourcesToAddOrRemove(AddOrRemove.Add, sourcesToAdd));
+        }
+    }
+
+    /**
+     * Queue a "source-remove" for remote sources, to be signaled once the session is established.
+     *
+     * @param sourcesToRemove the remote sources for the "source-remove".
+     */
+    public void queueRemoteSourcesToRemove(ConferenceSourceMap sourcesToRemove)
+    {
+        synchronized (queuedRemoteSourceChanges)
+        {
+            SourcesToAddOrRemove previous = Iterables.getLast(queuedRemoteSourceChanges, null);
+            if (previous != null && previous.getAction() == AddOrRemove.Remove)
+            {
+                // We merge sourcesToRemove with the previous sources queued to be remove to reduce the number of
+                // source-remove messages that need to be sent.
+                queuedRemoteSourceChanges.remove(queuedRemoteSourceChanges.size() - 1);
+                sourcesToRemove = sourcesToRemove.copy();
+                sourcesToRemove.add(previous.getSources());
+            }
+
+            queuedRemoteSourceChanges.add(new SourcesToAddOrRemove(AddOrRemove.Remove, sourcesToRemove));
+        }
+    }
+
+    /**
+     * Sets information about Colibri channels allocated for this participant.
+     *
+     * @param colibriChannelsInfo the IQ that holds colibri channels state.
+     */
+    public void setColibriChannelsInfo(ColibriConferenceIQ colibriChannelsInfo)
+    {
+        this.colibriChannelsInfo = colibriChannelsInfo;
+    }
+
+    /**
+     * Returns {@link ColibriConferenceIQ} that describes Colibri channels
+     * allocated for this participant.
+     */
+    public ColibriConferenceIQ getColibriChannelsInfo()
+    {
+        return colibriChannelsInfo;
+    }
+
+    /**
+     * Replaces the {@link ParticipantChannelAllocator}, which is currently
+     * allocating channels for this participant (if any) with the specified
+     * channel allocator (if any).
+     * @param channelAllocator the channel allocator to set, or {@code null}
+     * to clear it.
+     */
+    public void setChannelAllocator(ParticipantChannelAllocator channelAllocator)
+    {
+        synchronized (channelAllocatorSyncRoot)
+        {
+            if (this.channelAllocator != null)
+            {
+                // There is an ongoing thread allocating channels and sending
+                // an invite for this participant. Tell it to stop.
+                logger.warn("Canceling " + this.channelAllocator);
+                this.channelAllocator.cancel();
+            }
+
+            this.channelAllocator = channelAllocator;
+        }
+    }
+
+    /**
+     * Signals to this {@link Participant} that a specific
+     * {@link ParticipantChannelAllocator} has completed its task and its thread
+     * is about to terminate.
+     * @param channelAllocator the {@link ParticipantChannelAllocator} which has
+     * completed its task and its thread is about to terminate.
+     */
+    public void channelAllocatorCompleted(ParticipantChannelAllocator channelAllocator)
+    {
+        synchronized (channelAllocatorSyncRoot)
+        {
+            if (this.channelAllocator == channelAllocator)
+            {
+                this.channelAllocator = null;
+            }
+        }
     }
 
     /**
@@ -489,7 +680,6 @@ public class Participant
      *
      * @return
      */
-    @Override
     @NotNull
     public ConferenceSourceMap getSources()
     {
@@ -500,7 +690,6 @@ public class Participant
      * @return {@code true} if the Jingle session with this participant has
      * been established.
      */
-    @Override
     public boolean isSessionEstablished()
     {
         return jingleSession != null;
