+++ src/main/java/org/jitsi/jicofo/conference/colibri/ParticipantChannelAllocator.java
@@ -17,6 +17,7 @@
  */
 package org.jitsi.jicofo.conference.colibri;
 
+import org.checkerframework.checker.nullness.qual.*;
 import org.jitsi.impl.protocol.xmpp.*;
 import org.jitsi.jicofo.*;
 import org.jitsi.jicofo.codec.*;
@@ -37,6 +38,8 @@ import org.jxmpp.stringprep.*;
 
 import java.util.*;
 
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
+
 /**
  * An {@link AbstractChannelAllocator} which invites an actual participant to
  * the conference (as opposed to e.g. allocating Colibri channels for
@@ -46,7 +49,7 @@ import java.util.*;
  * @author Pawel Domas
  * @author Boris Grozev
  */
-public class ParticipantChannelAllocator extends AbstractChannelAllocator
+public class ParticipantChannelAllocator implements Runnable
 {
     /**
      * The constant value used as owner attribute value of
@@ -70,6 +73,47 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
     private final Logger logger;
 
     /**
+     * The {@link JitsiMeetConferenceImpl} into which a participant will be
+     * invited.
+     */
+    private final JitsiMeetConferenceImpl meetConference;
+
+    /**
+     * The {@link BridgeSession} on which
+     * to allocate channels for the participant.
+     */
+    private final BridgeSession bridgeSession;
+
+    /**
+     * A flag which indicates whether channel allocation is canceled. Raising
+     * this makes the allocation thread discontinue the allocation process and
+     * return.
+     */
+    private volatile boolean canceled = false;
+
+    /**
+     * First argument stands for "start audio muted" and the second one for
+     * "start video muted". The information is included as a custom extension in
+     * 'session-initiate' sent to the user.
+     */
+    private final boolean[] startMuted;
+
+    /**
+     * Indicates whether or not this task will be doing a "re-invite". It
+     * means that we're going to replace a previous conference which has failed.
+     * Channels are allocated on new JVB and peer is re-invited with
+     * 'transport-replace' Jingle action as opposed to 'session-initiate' in
+     * regular invite.
+     */
+    private final boolean reInvite;
+
+    /**
+     * The colibri channels that this allocator has allocated. They'll be
+     * cleaned up if the allocator is canceled or failed at any point.
+     */
+    private ColibriConferenceIQ colibriChannels;
+
+    /**
      * Override super's AbstractParticipant
      */
     private final Participant participant;
@@ -85,17 +129,216 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
             boolean reInvite,
             Logger parentLogger)
     {
-        super(meetConference, bridgeSession, participant, startMuted, reInvite, parentLogger);
+        this.meetConference = meetConference;
+        this.bridgeSession = bridgeSession;
+        this.startMuted = startMuted;
+        this.reInvite = reInvite;
         this.participant = participant;
         logger = parentLogger.createChildLogger(getClass().getName());
         logger.addContext("participant", participant.getChatMember().getName());
     }
 
     /**
-     * {@inheritDoc}
+     * Entry point for the {@link AbstractChannelAllocator} task.
      */
     @Override
-    protected Offer createOffer()
+    public void run()
+    {
+        try
+        {
+            doRun();
+        }
+        catch (Throwable e)
+        {
+            logger.error("Channel allocator failed: ", e);
+            cancel();
+        }
+        finally
+        {
+            if (canceled && colibriChannels != null)
+            {
+                bridgeSession.colibriConference.expireChannels(colibriChannels);
+            }
+
+            if (participant != null)
+            {
+                participant.channelAllocatorCompleted(this);
+            }
+        }
+    }
+
+    private void doRun()
+    {
+        Offer offer;
+
+        try
+        {
+            offer = createOffer();
+        }
+        catch (UnsupportedFeatureConfigurationException e)
+        {
+            logger.error("Error creating offer", e);
+            return;
+        }
+        if (canceled)
+        {
+            return;
+        }
+
+        colibriChannels = allocateChannels(offer.getContents());
+        if (canceled)
+        {
+            return;
+        }
+
+        if (colibriChannels == null)
+        {
+            logger.error("Channel allocator failed: " + participant);
+
+            // Cancel this task - nothing to be done after failure
+            cancel();
+            return;
+        }
+
+        if (participant != null)
+        {
+            participant.setColibriChannelsInfo(colibriChannels);
+        }
+
+        offer = updateOffer(offer, colibriChannels);
+        if (canceled)
+        {
+            return;
+        }
+
+        try
+        {
+            invite(offer);
+        }
+        catch (SmackException.NotConnectedException e)
+        {
+            logger.error("Failed to invite participant: ", e);
+        }
+    }
+
+    /**
+     * Allocates Colibri channels for this {@link AbstractChannelAllocator}'s
+     * {@link Participant} on {@link #bridgeSession}.
+     *
+     * @return a {@link ColibriConferenceIQ} which describes the allocated
+     * channels, or {@code null}.
+     */
+    private ColibriConferenceIQ allocateChannels(
+            List<ContentPacketExtension> contents)
+    {
+        Jid jvb = bridgeSession.bridge.getJid();
+        if (jvb == null)
+        {
+            logger.error("No bridge jid");
+            cancel();
+            return null;
+        }
+
+        // The bridge is faulty.
+        boolean faulty;
+        // We want to re-invite the participants in this conference.
+        boolean restartConference;
+        try
+        {
+            logger.info(
+                    "Using " + jvb + " to allocate channels for: "
+                            + (participant == null ? "null" : participant.toString()));
+
+            ColibriConferenceIQ colibriChannels = doAllocateChannels(contents);
+
+            // null means canceled, because colibriConference has been
+            // disposed by another thread
+            if (colibriChannels == null)
+            {
+                cancel();
+                return null;
+            }
+
+            bridgeSession.bridge.setIsOperational(true);
+            meetConference.colibriRequestSucceeded();
+            return colibriChannels;
+        }
+        catch (ConferenceNotFoundException e)
+        {
+            // The conference on the bridge has likely expired. We want to
+            // re-invite the conference participants, though the bridge is not
+            // faulty.
+            restartConference = true;
+            faulty = false;
+            logger.error(jvb + " - conference ID not found (expired?):" + e.getMessage());
+        }
+        catch (BadRequestException e)
+        {
+            // The bridge indicated that our request is invalid. This does not
+            // mean the bridge is faulty, and retrying will likely result
+            // in the same error.
+            // We observe this when an endpoint uses an ID not accepted by
+            // the new bridge (via a custom client).
+            restartConference = false;
+            faulty = false;
+            logger.error(
+                    jvb + " - the bridge indicated bad-request: " + e.getMessage());
+        }
+        catch (ColibriException e)
+        {
+            // All other errors indicate that the bridge is faulty: timeout,
+            // wrong response type, or something else.
+            restartConference = true;
+            faulty = true;
+            logger.error(
+                    jvb + " - failed to allocate channels, will consider the "
+                            + "bridge faulty: " + e.getMessage(), e);
+        }
+
+        // We only get here if we caught an exception.
+        if (faulty)
+        {
+            bridgeSession.bridge.setIsOperational(false);
+            bridgeSession.hasFailed = true;
+        }
+
+        cancel();
+
+        // If the ColibriConference is in use, and we want to retry,
+        // notify the JitsiMeetConference.
+        if (restartConference &&
+                isNotBlank(bridgeSession.colibriConference.getConferenceId()))
+        {
+            meetConference.channelAllocationFailed(jvb);
+        }
+
+        return null;
+    }
+
+    /**
+     * Raises the {@code canceled} flag, which causes the thread to not continue
+     * with the allocation process.
+     */
+    public void cancel()
+    {
+        canceled = true;
+    }
+
+    @Override
+    public String toString()
+    {
+        return String.format(
+                "%s[%s, %s]@%d",
+                this.getClass().getSimpleName(),
+                bridgeSession,
+                participant,
+                hashCode());
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    private Offer createOffer()
         throws UnsupportedFeatureConfigurationException
     {
         // Feature discovery
@@ -119,8 +362,7 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
     /**
      * {@inheritDoc}
      */
-    @Override
-    protected ColibriConferenceIQ doAllocateChannels(
+    private ColibriConferenceIQ doAllocateChannels(
         List<ContentPacketExtension> offer)
         throws ColibriException
     {
@@ -134,8 +376,7 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
     /**
      * {@inheritDoc}
      */
-    @Override
-    protected void invite(Offer offer)
+    private void invite(Offer offer)
         throws SmackException.NotConnectedException
     {
         /*
@@ -281,8 +522,7 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
     /**
      * {@inheritDoc}
      */
-    @Override
-    protected Offer updateOffer(Offer offer, ColibriConferenceIQ colibriChannels)
+    private @NonNull Offer updateOffer(Offer offer, ColibriConferenceIQ colibriChannels)
     {
         ConferenceSourceMap conferenceSources = meetConference.getSources()
                 .copy()
@@ -411,7 +651,6 @@ public class ParticipantChannelAllocator extends AbstractChannelAllocator
      * @return the {@link Participant} associated with this
      * {@link ParticipantChannelAllocator}.
      */
-    @Override
     public Participant getParticipant()
     {
         return participant;
